feat(materiaalaanvragen): move logic to the actual page
