--- src/modules/material-requests/components/MaterialRequestsOverview/index.ts
@@ -1,2 +1,0 @@
-export { default as MaterialRequestsOverview } from './MaterialRequestsOverview';
-export * from './MaterialRequestsOverview.types';
