--- src/modules/material-requests/components/index.ts
@@ -1,1 +1,0 @@
-export * from './MaterialRequestsOverview';
